% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_data.R
\name{classify}
\alias{classify}
\title{This function allows you to build the needed list for the API to classify the text.}
\usage{
classify(api_token = NULL, query = NULL, endpoint = "classify")
}
\arguments{
\item{api_token}{string of the  token from your personal API access}

\item{query}{query of texts to classify in form of a list generated by query_maker().}

\item{endpoint}{character string with the name of the endpoint}
}
\value{
a vector with predictions
}
\description{
This function allows you to build the needed list for the API to classify the text.
}
\examples{
\dontrun{
predictions <- classify(api_token = "", endpoint = "classify", query = query_list)
}
}
