% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_data.R
\name{classify_data}
\alias{classify_data}
\title{This function allows you to build the needed list for the API to classify the text.}
\usage{
classify_data(
  api_token = NULL,
  endpoint = "classify",
  query = NULL,
  data = NULL,
  text_name = NULL
)
}
\arguments{
\item{api_token}{string of the  token from your personal API access}

\item{endpoint}{character string with the name of the endpoint}

\item{query}{query of texts to classify in form of a list generated by query_maker().}

\item{data}{data frame with the text variable to classify}

\item{text_name}{character string with the name of the text variable to classify.}
}
\value{
a vector with predictions
}
\description{
This function allows you to build the needed list for the API to classify the text.
}
\examples{
\dontrun{
predictions <- classify_data(api_token = "", endpoint = "classify", data = df)
}
}
