% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{This function allows you to load data with text that you want to classify}
\usage{
load_data(input)
}
\arguments{
\item{input}{path to data of type csv, xlsx or rds.}
}
\value{
a data frame
}
\description{
This function allows you to load data with text that you want to classify
}
\examples{
\dontrun{
df <- load_data(input = "./DATA.csv")
}
}
