% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_maker.R
\name{query_maker}
\alias{query_maker}
\title{This function allows you to build the needed list for the API to classify the text.}
\usage{
query_maker(data = NULL, text_name = NULL)
}
\arguments{
\item{data}{data frame with the text column to classify.}

\item{text_name}{character string with the name of the text variable to classify.}
}
\value{
a list
}
\description{
This function allows you to build the needed list for the API to classify the text.
}
\examples{
\dontrun{
dt <- data.frame(data.frame(id = c (1:5),
                            text = c("This","is","a","Test","data frame"))

df <- query_maker(data = dt, text_name = "text")
}
}
