% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_data}
\alias{write_data}
\title{This function allows you to write data with text that you classified with the API}
\usage{
write_data(output, data)
}
\arguments{
\item{output}{path to data of type csv, xlsx or rds.}

\item{data}{data frame to write to disk.}
}
\description{
This function allows you to write data with text that you classified with the API
}
\examples{
\dontrun{
write_data(output = "./DATA.csv", data = df)
}
}
